Correctly process drag start
The egui drag event has changed and now the drag only starts after
the pointer has moved. This means that the pointer could move to
a different node before the drag starts and the wrong node is dragged.

To fix this, the drag state has to be set when the pointer is pressed
regardless if it is a correct drag or not.
We already do our own drag validation so the egui drag event
no longer matters to us.