Fix incorrect behavior when dragging and clicking
1) Right clicks should not active a context menu while a drag is valid.
If a drag is not yet valid and a right click occurs, the drag should be
canceled so that only a drag or a context menu is valid at any given
time; never both.
The fix for this is simply checking if a drag is valid before
registering the right click action.
Once a right click action is registered, any drag actions should be
cleared.

2) A drag should only be completed by releasing the primary pointer
button.

3) If during a drag a secondary button is pressed, the hover position
of the egui::Response has the hover position set to None.
To avoid this we can simply use the lastest pointer position.