Split TreeViewData and TreeViewState
The long term goal is it to draw the entire tree and then handle input
for the entire tree at once. At the moment input is handled partly
when a node is added to the tree and partly after the whole tree
is constructed.
The old approach doesn't  really work with multiple selections since
it would sometimes be necessary to select a node that has not yet
been added to the tree.

Handling all input after the tree is constructed will allow me to
always handle input with the correct tree.

The goal of this commit is to eliminate much of TreeViewData since
it is mostly used to allow for the handling of input when a node is
constructed. If all input is handled at the end, then TreeViewData
is not necessary in its current form.