Correctly handle context menus for multi selection
When a context menu is request on a node that is part of a multi
selection then that context menu should be relevant to all selected
nodes; not just the node that was right clicked. This means we cannot
show a context menu that was configure for any particular node since
that menu is only supposed to be relevant to the node it was configured
for.

Instead we show the new fall back context menu. This also serves
as a fall back if a node did not configure a context menu.

If you wanted to, you could configure all your context menus in the tree
to use the fall back context menu. Whatever fits better.