Fix panic caused by nonexistent node in selection
If a node is deleted from the tree but it was previously selected
the tree wants to access this nodes rectangle to draw its background.
This will cause a panic because the node doesn't have a background
since its not part of the tree anymore.

To fix this we have to make sure that the list of selected nodes only
contains nodes that actually exists in the tree. I make the node_states
private and provide a setter method to set the states and only retain
existing nodes in the selection.