Refactor a bit
We dont need to save the modifiers to make the activate action
happen, instead we can just take a boolean to show that the
action should happen and then read the modifiers there.

The `TreeViewState` should not handle the key press for the
activate action. We can move this to the `handle_input` function
instead.