Fix double clicking behavior
Since the entire tree is a single ui, egui thinks the tree was double
clicked even if two different nodes were clicked.
To know for sure that a single node was double clicked we simply
have to remember which node was the last that was clicked.
A double click is only valid if the same node was clicked previously