Don't draw non visible nodes
Add a node height to nodes. Using the height we can calculate if a node
would be visible in the current clip rect.
If no node height is provided by the user the tree uses a default node
height. The default is the interaction height out of
egui::style::Spacing. This is also the default height for a Ui::horizontal
group and should work for all standard label based nodes.
Not drawing non-visible nodes improves performance for big trees.