Make the indent hint work without the closer rect
If a dir is not visible we still want to show its indent hint. For this
we cant use the closer rect because the dir might not have one if
it is not visible. Instead we store the row range (which is calculated
for every node, always) and calculate the correct indent hint
position from that.