Remove the requirement for Copy of NodeId
NodeIds should be small and cheap to clone. Copy however has the added
requirement that it must be a simple mem copy to clone the value. This
disqualifies NodeIds that contain a RC for example. Since RC's are never
Copy you could not use an RC as a node id even if they are cheap to
clone.