Initialize the tree view with the tree
When the openness hashmap is empty the lookup is practically instant.
Although we get great performance in that case, will degrade over time
as the openness hashmap slowly becomes filled up. This change
fills the hashmap will data for each directory and simulates the worst
case scenario for further profiling.