Fix cargo features for playground example
The playground example uses the Uuid crate for unique ids. When the
persistence feature is active, the serde feature of the uuid crate must
also be active for the example to compile correctly.

To be honest i dont love that the uuid/serde feature is now required
for the persistence feature. Atleast in the way it is configured right
now. Uuid is a dev dependency and i would like to keep is seperate
from the lib dependencies if i can. I dont think thats possible
so this looks like the nest best thing.