FIx interaction when inside a scroll area
The tree view gets its height from the available height of the
current ui. Inside a scroll area you would expect the available
height to be the height of the scrollable content. This is not the
case. Instead it is the height of the viewport of the scroll area.

This caused the interaction rectangle to not be tall enough
and scroll off screen in a scroll area. As a result, nodes
below the viewport were not clickable.

The fix to give the interaction rectangle an explicit height.
The only logical choice for the height is the height of the last
frame. This fixes the issue as far as i can tell.