Scroll to the visible node when in a scroll area
When using the keyboard to select a node that is just outside of the
visible area the scroll area should scroll to make that node visible.
The scroll area should scroll to make the cursor node visible since
that is the node that is most likely to be the one that needs to be
visible.

To do this we have to split the input handling into two phases.
The first phase can be done when the node is structually visible.
Here the keyboard navigation can take place since we dont actually
need to see the node to be able to navigate to it.
Plus since a node that is just outside the scroll area is not visible
we have to be able to navigate to structually visible nodes.

Phase two can only happen if the node is actually visible.
Here clicking and dragging needs to happen since these
inputs can only function on nodes that are visible.