Fix flickering on context menus
When a node is right clicked while a context menu of another node was
still open it would sometimes flash the "id already in use" warning.
This happens because the tree would show the context menu for the
currently secondary selected node and the node that was freshly clicked.
We should instead be a little smarter when to show the context menu
to avoid this double drawing.

A second reason why the context menus flicker is because a node was
right clicked while a context menu was already open on a previous node.
When we draw the context menu for the previous node we don't now yet
if a later node is right clicked or not. When it turns out that the next
node was right clicked we would draw the new context menu on top.
This had the possibility of creating flickering again.
The fix here is to not render the next context menu if a previous
context menu was already drawn.
The context menu will move to the new spot but will still be rendered
with the old node. As far as i can tell there is no good way around this
except for custom context menus.