Fix drag output is double with drag external
The drag and move external action should only be output if the
drag does not find a valid target within the tree.
The external actions should also not be output if the drop
is over the currently dragged node. Such a drop is not a valid
drop target but it is still considered an internal drag and drop

Fixes issue #28