Avoid duplicates in the simplified drag vector
When dragging multiple nodes the simplified drag vector some nodes
multiple times. This was because the DragStarted input did not terminate
itself.

The fix is to make the simplified drag logic more robust.

This also removes a match on the output to determine if the current
branch is dragged. Im not entirely sure if this change is correct but
it seams to work sofar.