Fix crash when dragging a node outside the window
I am not 100% sure what is happening here but if drag a node and drop
it outside of the window very quickly then `latest_pos` might be None
which causes a panic later on.
Egui already filters out the None value when using `interact_pos` which
avoids this problem.

Relates to #35