Fix node flatten bug
A node that was flattened was not visible at all.
The problem was that the open state of the directory was set to false
if a node was flattened. This caused all its children to not be visible.

The fix here is to set the open state for flattened nodes to open. The
flattened node must remain not visible though.